plugins/plugin-vision/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "elizaos-plugin-vision"
version = "2.0.0"
description = "Vision plugin for elizaOS - provides camera integration and visual awareness"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "elizaOS Contributors" },
]
keywords = [
    "plugin",
    "elizaos",
    "vision",
    "camera",
    "computer-vision",
    "ai-vision",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-xprocess>=0.23.0,<1.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
ocr = [
    "pytesseract>=0.3.10",
]
cv = [
    "opencv-python>=4.8.0",
]

[project.urls]
Homepage = "https://elizaos.ai"
Repository = "https://github.com/elizaos/elizaos"
Documentation = "https://docs.elizaos.ai"

[tool.hatch.build.targets.wheel]
packages = ["elizaos_vision"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v -p no:anchorpy"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

